requirements.txt
streamlit==1.32.0
pandas==2.2.1
Pillow==10.2.0
python-dotenv==1.0.1
bcrypt==4.1.2
